Create mrs_tasks.py module; remove unneeded imports
